factored out dqn class and parameters
